+++ src/Entity/Metadata/MetadataContainerAbstract.php
@@ -57,7 +57,7 @@ abstract class MetadataContainerAbstract extends CollectionAbstract
     {
         $data = [];
 
-        if (is_array($metas)) {
+        if (\is_array($metas)) {
             foreach ($metas as $meta) {
                 $data[$meta['key']] = $meta['value'];
             }
@@ -71,7 +71,7 @@ abstract class MetadataContainerAbstract extends CollectionAbstract
         if ($data instanceof CollectionAbstract) {
             return $data->get($piece);
         }
-        if (is_array($data) && array_key_exists($piece, $data)) {
+        if (\is_array($data) && array_key_exists($piece, $data)) {
             return $data[$piece];
         }