+++ src/Console/AbstractApplication.php
@@ -19,7 +19,10 @@ namespace Gpupo\CommonSdk\Console;
 
 use Gpupo\Common\Console\AbstractApplication as Core;
 use Gpupo\CommonSchema\TranslatorDataCollection;
+use Gpupo\CommonSdk\FactoryInterface;
 use Monolog\Logger;
+use Psr\Log\LoggerInterface;
+use Psr\SimpleCache\CacheInterface;
 use Symfony\Component\Cache\Simple\FilesystemCache;
 use Symfony\Component\Console\Input\ArgvInput;
 use Symfony\Component\Console\Output\ConsoleOutput;
@@ -29,6 +32,8 @@ use Symfony\Component\Yaml\Yaml;
 
 abstract class AbstractApplication extends Core
 {
+    abstract public function factorySdk(array $options, LoggerInterface $logger = null, CacheInterface $cache = null): FactoryInterface;
+
     public function displayOrderList(TranslatorDataCollection $collection, OutputInterface $output)
     {
         if (0 === $collection->count()) {
@@ -43,22 +48,21 @@ abstract class AbstractApplication extends Core
 
     public function init(string $namespace, string $rootDirectory)
     {
-        $cache = new FilesystemCache();
         $input = new ArgvInput();
         $output = new ConsoleOutput();
-
-        $logger = new Logger('console');
         $config = getenv();
 
-        $localFilename = sprintf('%s/%s', $rootDirectory, $config['extra_file']);
-        if (file_exists($localFilename)) {
-            $data = Yaml::parseFile($localFilename);
-            if (\is_array($data)) {
-                $config = array_merge($config, $data);
+        if (!empty($config['extra_file'])) {
+            $localFilename = sprintf('%s/%s', $rootDirectory, $config['extra_file']);
+            if (file_exists($localFilename)) {
+                $data = Yaml::parseFile($localFilename);
+                if (\is_array($data)) {
+                    $config = array_merge($config, $data);
+                }
             }
         }
 
-        $factory = $this->factorySdk($config, $logger, true, $cache);
+        $factory = $this->factorySdk($config, new Logger('console'), new FilesystemCache());
 
         $finder = new Finder();
         $finder->files()->name('*Command.php')->notName('*Abstract*')->in(sprintf('%s/src/Console/Command', $rootDirectory));
@@ -66,7 +70,7 @@ abstract class AbstractApplication extends Core
         foreach ($finder as $file) {
             $class = str_replace('.php', '', $file->getRelativePathname());
             $segments = explode('/', $class);
-            $class = $namespace.'\\'.implode('\\', $segments);
+            $class = $namespace.'\\Command\\'.implode('\\', $segments);
             $this->add(new $class($factory));
         }